+++ kubejs/server_scripts/base/tags/items/forge/stripped_logs.js
@@ -1,0 +1,3 @@
+ServerEvents.tags('item', (event) => {
+    event.add('forge:stripped_logs', [/log_stripped/, /stripped_.*)(log|stem|pedu)/]);
+});