+++ kubejs/server_scripts/base/tags/items/forge/stripped_wood.js
@@ -1,0 +1,7 @@
+ServerEvents.tags('item', (event) => {
+    event.add('forge:stripped_wood', [
+        /wood_stripped/,
+        /stripped_.*_(wood|hyphae)/,
+        'ars_elemental:stripped_yellow_archwood'
+    ]);
+});