+++ changelogs/CHANGELOG.md
@@ -15,6 +15,7 @@
 -   Correct quest description for bookwyrm familiars [\#912](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/912)
 -   Fixed an issue with Gateways crashing the game if used in a Compact Machine [\#912](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/912)
 -   [Expert] Mekanism tanks may no longer be upgraded in world [\#912](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/912)
+-   Fix missing stripped log/wood tags [\#912](https://github.com/EnigmaticaModpacks/Enigmatica9/issues/912)
 
 ### Enigmatica 9 v1.20.1